Fix orientation and alignment being ignored
By effectively calling `$(...).val(default_value)` on the group of radio
buttons, the previous implementation overwrote the value of each button
with the same value. With each button in the group having the same
value, the selection in the UI had no effect.

This commit fixes the issue by wrapping the default value into an array.
If the argument to the `.val` function is an array, jQuery no-longer
overwrites the values of the elements, but instead sets or unsets their
`checked` and `selected` attributes depending on whether the value of
the element is inside the array.